refactored component and container locations
