--- components/DefaultCurrency.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default () => (
-  <div>Default Currency Selector</div>
-)
-
-// Dispatches action to change default currency and get new Prices List
