--- components/Prices.tsx
@@ -1,5 +1,0 @@
-import React from 'react'
-
-export default () => (
-  <div>Price List</div>
-)
