+++ components/index.ts
@@ -1,0 +1,6 @@
+export { default as CurrencySelector } from './CurrencySelector/CurrencySelector';
+export { default as Header } from './Header/Header';
+export { default as Prices } from './Prices/Prices';
+export { default as Navigation } from './Navigation/Navigation';
+export { default as wallets } from './Wallets/Wallets';
+export { default as wallet } from './Wallets/Wallet';
