+++ pages/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import FiatWallet from '../components/FiatWallet'
+import FiatWallet from '../containers/FiatWallet'
 
 export default () => (
   <FiatWallet />
